requirements.txt
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10.3
django-extensions==1.7.4
django-honeypot==0.5.0
gunicorn==19.4.1
psycopg2==2.6.1
python-decouple==3.0
static3==0.7.0
django-test-without-migrations==0.4
pytz
django-localflavor==1.3
